update empathy map
Additional points added to the empathy maps of the relationship manager and the call management center. In terms of the Call management center i have tried to consider the center as a group of relationship managers and their managers. Together as a team how do they think of these challenges i think is different to how each individual relationship manager might think.